+++ src/client/js/functions.js
@@ -1,5 +1,8 @@
 /*
  * pwix:tabular/src/client/js/functions.js
+ *
+ * Note that we cannot really reluy on 'init.dt' end of DataTable initialization event as it may be triggered early enough in the process
+ * And the TabularExt template may not be yet rendered. So Have tio use initComplete() DataTable option which is also used by 'aldeed:tabular'
  */
 
 import { check, Match } from 'meteor/check';
@@ -9,34 +12,27 @@ const logger = Logger.get();
 
 /**
  * @param {String} tabularName the tabular name
- * @param {String} selector the parent selector to be sure to intercept the initialization message
- * @param {Object} opts an optional options object, with follwoing keys:
- *  - first: whether to wait for init message, defaulting to false
+ * @param {dtApi} dtTable the DataTable dtApi entry point
  */
-Tabular.applyColumnsState = function( tabularName, selector, opts={} ){
+Tabular.applyState = function( tabularName, dtTable ){
     check( tabularName, Match.NonEmptyString );
-    check( selector, Match.NonEmptyString );
 
     // apply state from settings (if any)
-    const _applyState = function(){
-        const $table = $( selector ).find( 'table.dataTable' );
-        const dtTable = $table.DataTable();
-        const tabularTable = Package['aldeed:tabular'].default.tablesByName[tabularName];
+    // order and visibility of the columns
+    const _applyColumnsState = function( dtTable, tabularTable ){
         const columns = tabularTable.options.columns;
         const indexMap = Tabular.indexMap( columns );
         const settings = Tabular.getSettingsColumns( tabularName );
-        logger.debug( 'columns', columns, 'settings', settings );
         if( settings ){
             const allNames = columns.map(( it ) => it.name );
             // apply ordering
             const otherColumns = columns.filter(( it ) => !settings.includes( it.name ));
-            logger.debug( 'otherColumns', otherColumns );
             const order = settings.map(( name ) => indexMap[name] );
             for( const it of otherColumns ){
                 order.push( it.index );
             }
-            logger.debug( 'order', order );
             dtTable.colReorder.order( order );
+            //logger.debug( 'columns', columns, 'settings', settings, 'order', order );
             // apply visibility
             for( const it of columns ){
                 const visible = settings.includes( it.name )
@@ -47,14 +43,14 @@ Tabular.applyColumnsState = function( tabularName, selector, opts={} ){
         }
     };
 
-    if( opts.first ){
-        const $parent = $( selector );
-        $parent.one( 'init.dt', function(){
-            _applyState();
-        });
-    } else {
-        _applyState();
-    }
+    // apply state from settings (if any)
+    const _applyRowsState = function( dtTable, tabularTable ){
+    };
+
+    const tabularTable = Package['aldeed:tabular'].default.tablesByName[tabularName];
+    //logger.debug( 'tabularName', tabularName, 'dtTable', dtTable, 'tabularTable', tabularTable );
+    _applyColumnsState( dtTable, tabularTable );
+    _applyRowsState( dtTable, tabularTable );
 }
 
 /**